+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Version 2.3 (not yet released)
 
+* Fixed issue with Event and Participant Tokens.
+
 ## Version 2.2
 
 * Fixed issue with Send to Related contact action.